Se arreglaron los paths de los sing ups, ademas de acomodar las pantallas en sus respetivas carpetas
