+++ lib/screens/screens.dart
@@ -1,3 +1,8 @@
 export 'app.dart';
 export 'home.dart';
-export 'SingUp4.dart';
+
+//Export signup 4
+export 'singup4/singup4.dart';
+
+//Export signup 1
+export 'singup1/singup1.dart';
