+++ lib/screens/singup4/singup4.dart
@@ -1,5 +1,5 @@
 import 'package:auto_matic/config/config.dart';
-import 'package:flutter/material.dart';
+import 'package:flutter/cupertino.dart';
 
 void main() {
   runApp(MyApp());
@@ -140,21 +140,21 @@ class _AccessWithSocialMediaSection extends StatelessWidget {
   Widget build(BuildContext context) {
     return Column(
       children: [
-        Text('Accede con', style: TextStyle(fontSize: 18)),
+        Text('Accede con', style: TextStyle(fontSize: 24,fontWeight: FontWeight.bold, color: Config.secondColor)),
         SizedBox(height: 10),
         Row(
           mainAxisAlignment: MainAxisAlignment.center,
           children: [
             Container(
-              width: 50,
-              height: 50,
+              width: 60,
+              height: 60,
               color: Colors.blue,
               child: Center(child: Text('Redes Sociales')),
             ),
             SizedBox(width: 20),
             Container(
-              width: 50,
-              height: 50,
+              width: 60,
+              height: 60,
               color: Colors.green,
               child: Center(child: Text('Google')),
             ),
