+++ lib/widgets/WidgetTimeline.dart
@@ -1,65 +1,4 @@
 import 'package:auto_matic/config/config.dart';
-
-
-void main() {
-  runApp(MyApp());
-}
-
-class MyApp extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      debugShowCheckedModeBanner: false,
-      home: Scaffold(
-        body: MyBody(),
-      ),
-    );
-  }
-}
-
-class MyBody extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      color: Config.firstColor, // Fondo blanco
-      padding: EdgeInsets.all(16.0),
-      child: Column(
-        mainAxisAlignment: MainAxisAlignment.start, // Alinear al principio
-        crossAxisAlignment: CrossAxisAlignment.start, // Alinear a la izquierda
-        children: [
-          Align( // Alinear el TextField a la esquina superior izquierda
-            alignment: Alignment.topLeft,
-            child: ConstrainedBox( // Limitar el ancho del TextField
-              constraints: BoxConstraints(maxWidth: MediaQuery.of(context).size.width * 0.6),
-              child: TextField(
-                decoration: InputDecoration(
-                  labelText: 'Introduce el código único del producto',
-                ),
-              ),
-            ),
-          ),
-          SizedBox(height: 16.0),
-          ElevatedButton(
-            onPressed: () {
-              // Lógica para el botón "Continuar"
-            },
-            child: Text('Continuar', style: TextStyle(color: Colors.white)), // Letra blanca
-            style: ElevatedButton.styleFrom(
-              backgroundColor: Color(0xFF208454), // Color hexadecimal #208454
-            ),
-          ),
-          SizedBox(height: 32.0),
-          TimelineWidget(), // Línea de tiempo
-          SizedBox(height: 16.0),
-
-          SizedBox(height: 16.0),
-          Text('Nombre de la empresa'), // Nombre de la empresa
-        ],
-      ),
-    );
-  }
-}
-
 class TimelineWidget extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
